requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
black==23.12.1
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
crispy-bootstrap4==2023.1
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.7
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-markdownify==0.9.3
django-redis==5.4.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.0.3
h11==0.14.0
idna==3.6
inflection==0.5.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.4
locust==2.24.1
Markdown==3.5.1
MarkupSafe==2.1.5
msgpack==1.0.8
mypy-extensions==1.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gettext==5.0
python3-openid==3.2.0
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.3.1
roundrobin==0.0.4
rpds-py==0.18.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tinycss2==1.2.1
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uWSGI==2.0.23
vine==5.1.0
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.2
